déplace typeCaseEnum dans la classe Case
