+++ view/AccueilPanel.java
@@ -17,6 +17,7 @@ import javax.swing.JTextArea;
 import javax.swing.UIManager;
 
 import controller.AccueilControleur;
+import model.Case.typeCaseEnum;
 import model.Partie;
 import model.Partie.bonusEnum;
 import model.Stream;
@@ -116,11 +117,11 @@ public class AccueilPanel extends JPanel {
         panel.add(label);
 
         // Affiche chaque élément du model présent dans le jeu + description
-        for (Enum<?> element : type.equals("Bonus") ? Partie.bonusEnum.values() : Partie.typeCaseEnum.values()) {
-            String description = element instanceof Partie.bonusEnum ? ((Partie.bonusEnum) element).getDescription() : ((Partie.typeCaseEnum) element).getDescription();
+        for (Enum<?> element : type.equals("Bonus") ? Partie.bonusEnum.values() : typeCaseEnum.values()) {
+            String description = element instanceof Partie.bonusEnum ? ((Partie.bonusEnum) element).getDescription() : ((typeCaseEnum) element).getDescription();
             if(description != null) {
                 // Récupération du chemin de l'image
-                String pathImage = element instanceof Partie.bonusEnum ? ((Partie.bonusEnum) element).getPathImage() : ((Partie.typeCaseEnum) element).getPathImage();
+                String pathImage = element instanceof Partie.bonusEnum ? ((Partie.bonusEnum) element).getPathImage() : ((typeCaseEnum) element).getPathImage();
 
                 // Création d'un panel pour chaque élément
                 JPanel subPanel = new JPanel();
